fix tests before move
