+++ core/test/fraggle-rock/gather/base-artifacts-test.js
@@ -4,8 +4,8 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {getBaseArtifacts, finalizeArtifacts} from '../../../fraggle-rock/gather/base-artifacts.js';
-import {initializeConfig} from '../../../fraggle-rock/config/config.js';
+import {getBaseArtifacts, finalizeArtifacts} from '../../../gather/base-artifacts.js';
+import {initializeConfig} from '../../../config/config.js';
 import {createMockDriver} from './mock-driver.js';
 import {LighthouseError} from '../../../lib/lh-error.js';
