+++ core/test/fraggle-rock/gather/driver-test.js
@@ -4,7 +4,7 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {Driver} from '../../../fraggle-rock/gather/driver.js';
+import {Driver} from '../../../gather/driver.js';
 import {fnAny} from '../../test-utils.js';
 import {createMockCdpSession} from './mock-driver.js';
