+++ core/test/fraggle-rock/gather/navigation-runner-test.js
@@ -12,7 +12,7 @@ import {
   mockDriverSubmodules,
   mockRunnerModule,
 } from './mock-driver.js';
-import {initializeConfig} from '../../../fraggle-rock/config/config.js';
+import {initializeConfig} from '../../../config/config.js';
 import {defaultNavigationConfig} from '../../../config/constants.js';
 import {LighthouseError} from '../../../lib/lh-error.js';
 import DevtoolsLogGatherer from '../../../gather/gatherers/devtools-log.js';
@@ -24,7 +24,7 @@ import {Runner as runnerActual} from '../../../runner.js';
 // Some imports needs to be done dynamically, so that their dependencies will be mocked.
 // See: https://jestjs.io/docs/ecmascript-modules#differences-between-esm-and-commonjs
 //      https://github.com/facebook/jest/issues/10025
-/** @type {import('../../../fraggle-rock/gather/navigation-runner.js')} */
+/** @type {import('../../../gather/navigation-runner.js')} */
 let runner;
 
 const mocks = await mockDriverSubmodules();
@@ -33,7 +33,7 @@ beforeEach(async () => {
   mockRunner.reset();
   mockRunner.getGathererList.mockImplementation(runnerActual.getGathererList);
   mockRunner.getAuditList.mockImplementation(runnerActual.getAuditList);
-  runner = (await import('../../../fraggle-rock/gather/navigation-runner.js'));
+  runner = (await import('../../../gather/navigation-runner.js'));
 });
 
 /** @typedef {{meta: LH.Gatherer.GathererMeta<'Accessibility'>, getArtifact: Mock<any, any>, startInstrumentation: Mock<any, any>, stopInstrumentation: Mock<any, any>, startSensitiveInstrumentation: Mock<any, any>, stopSensitiveInstrumentation:  Mock<any, any>}} MockGatherer */
@@ -44,7 +44,7 @@ describe('NavigationRunner', () => {
   let requestor;
   /** @type {ReturnType<typeof createMockDriver>} */
   let mockDriver;
-  /** @type {import('../../../fraggle-rock/gather/driver.js').Driver} */
+  /** @type {import('../../../gather/driver.js').Driver} */
   let driver;
   /** @type {LH.Puppeteer.Page} */
   let page;
