+++ docs/user-flows.md
@@ -40,7 +40,7 @@ In DevTools, navigation is easy: ensure it's the selected mode and then click _A
 ```js
 import {writeFileSync} from 'fs';
 import puppeteer from 'puppeteer';
-import lighthouse from 'lighthouse/core/fraggle-rock/api.js';
+import lighthouse from 'lighthouse/core/api.js';
 
 (async function() {
   const browser = await puppeteer.launch();
@@ -90,7 +90,7 @@ In DevTools, select "Timespan" as the mode and click _Start timespan_. Record wh
 ```js
 import {writeFileSync} from 'fs';
 import puppeteer from 'puppeteer';
-import lighthouse from 'lighthouse/core/fraggle-rock/api.js';
+import lighthouse from 'lighthouse/core/api.js';
 
 (async function() {
   const browser = await puppeteer.launch();
@@ -125,7 +125,7 @@ In DevTools, select "Snapshot" as the mode. Set up the page in the state you wan
 ```js
 import {writeFileSync} from 'fs';
 import puppeteer from 'puppeteer';
-import lighthouse from 'lighthouse/core/fraggle-rock/api.js';
+import lighthouse from 'lighthouse/core/api.js';
 
 (async function() {
   const browser = await puppeteer.launch();
@@ -164,7 +164,7 @@ The below example codifies a user flow for an ecommerce site where the user navi
 import {writeFileSync} from 'fs';
 import puppeteer from 'puppeteer';
 import * as pptrTestingLibrary from 'pptr-testing-library';
-import lighthouse from 'lighthouse/core/fraggle-rock/api.js';
+import lighthouse from 'lighthouse/core/api.js';
 
 const {getDocument, queries} = pptrTestingLibrary;
